+++ src/engine/include/engine/map.h
@@ -32,7 +32,7 @@ class MapLayer {
 public:
     MapLayer(Types::Map2D data, std::shared_ptr<Tileset> tileset, uint32_t width, uint32_t height);
 
-    bool animate(float frameDiff);
+    bool animate(uint64_t frameDiff);
 
     void draw(Renderer& renderer, const Types::Rect<>& dst, bool clip);
     void drawRow(Renderer& renderer, const Types::Rect<>& dst, int row, TilesetAttribute::Type type, bool clip);
@@ -66,7 +66,7 @@ class Map {
 public:
     Map(const std::string& name);
 
-    bool animate(float frameDiff);
+    bool animate(uint64_t frameDiff);
 
     void draw(Renderer& renderer, const Types::Rect<>& dst, bool clip = true);
     void drawRow(Renderer& renderer, const Types::Rect<>& dst, int row, TilesetAttribute::Type type, bool clip = true);
